requirements.txt
openai
pyttsx3
SpeechRecognition
python-dotenv
